Refactored code so controllers are split into seperate files from main.py.
