+++ app/config.py
@@ -1,0 +1,2 @@
+save_path_json = 'uploads/json/'
+save_path_img = 'uploads/img/'
