+++ app/modules/auth.py
@@ -1,0 +1,21 @@
+import secrets
+
+from fastapi import FastAPI, Depends, HTTPException
+from fastapi.security import HTTPBasic, HTTPBasicCredentials
+from starlette import status
+
+app = FastAPI()
+
+security = HTTPBasic()
+
+
+def authorize(credentials: HTTPBasicCredentials = Depends(security)):
+    correct_username = secrets.compare_digest(credentials.username, "drone")
+    correct_password = secrets.compare_digest(credentials.password, "drone")
+    if not (correct_username and correct_password):
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Incorrect email or password",
+            headers={"WWW-Authenticate": "Basic"},
+        )
+    return credentials.username
